refactureing txpowerfix

+++ gluon-txpowerfix/files/lib/gluon/txpowerfix/txpowerfix.lua
@@ -3,7 +3,6 @@
 site = require("gluon.site")
 local uci = require("simple-uci").cursor()
 
-
 --- wrapper for calling systemcommands
 function cmd(_command)
         local f = io.popen(_command)
@@ -89,41 +88,13 @@ if interface24 and interface50 then
         end
 end
 
---- 2.4G
---- set values (1st pass)
+--- set HT-modes (1st pass)
 if interface24 then
 	uci:set('wireless', interface24, 'country', country)
 	uci:set('wireless', interface24, 'htmode', 'HT20')
 	uci:save('wireless')
 	uci:commit('wireless')
-	t = cmd('/sbin/wifi reconfigure')
-	t = cmd('/sbin/wifi down')
-	t = cmd('kill -9  $(ps |grep /usr/sbin/hostapd|grep -v grep|awk \'{print $1}\')')
-	t = cmd('killall hostapd >/dev/null 2>&1')
-	t = cmd('rm -f /var/run/wifi-*.pid >/dev/null 2>&1')
-	t = cmd('/sbin/wifi up')
-	t = cmd('sleep 5')
---- get maximum available power and step
-        t = cmd('iwinfo ' .. interface24 .. ' txpowerlist | tail -n 2 | head -n 1 | awk \'{print $1}\'')
-        maximumTxPowerDb = string.gsub(t, "\n", "")
-        maximumTxPowerDb = tonumber(maximumTxPowerDb)
-
-        if maximumTxPowerDb < 30 then
-                t = cmd('iwinfo ' .. interface24 .. ' txpowerlist | wc -l')
-                maximumTxPower = string.gsub(t, "\n", "")
-                maximumTxPower = tonumber(maximumTxPower)-1
-        else
-                t = cmd('iwinfo ' .. interface24 .. ' txpowerlist | grep -n "19 dBm" | cut -f1 -d\':\'')
-                maximumTxPower = string.gsub(t, "\n", "")
-                maximumTxPower = tonumber(maximumTxPower)-1
-        end
---- set values (2nd pass)
-        uci:set('wireless', interface24, 'txpower', maximumTxPower)
-        uci:save('wireless')
-        uci:commit('wireless')
 end
-
---- 5G
 if interface50 then
         if interface50 == 'radio0' or interface50 == 'radio1' then
 		if channel50 ~= 'auto' and uci:get('gluon', 'wireless', 'outdoor') ~= '1' then  --- do't do if outdoor is enabled
@@ -132,39 +103,51 @@ if interface50 then
 	                if string.match(VHT, 'HT') then
         	                uci:set('wireless', interface50, 'htmode', VHT)
                 	end
-			--- set values (1st pass)
-       	        	uci:save('wireless')
-               		uci:commit('wireless')
-               		t = cmd('/sbin/wifi reconfigure')
-               		t = cmd('/sbin/wifi down')
-			t = cmd('kill -9  $(ps |grep /usr/sbin/hostapd|grep -v grep|awk \'{print $1}\')')
-		        t = cmd('killall hostapd >/dev/null 2>&1')
-			t = cmd('rm -f /var/run/wifi-*.pid >/dev/null 2>&1')
-			t = cmd('/sbin/wifi up')
-			t = cmd('sleep 5')
-                	--- get maximum available power and step
-                	t = cmd('iwinfo ' .. interface50 .. ' txpowerlist | tail -n 2 | head -n 1 | awk \'{print $1}\'')
-                	maximumTxPowerDb = string.gsub(t, "\n", "")
-                	maximumTxPowerDb = tonumber(maximumTxPowerDb)
-                	if maximumTxPowerDb < 30 then
-                        	t = cmd('iwinfo ' .. interface50 .. ' txpowerlist | wc -l')
-                        	maximumTxPower = string.gsub(t, "\n", "")
-                        	maximumTxPower = tonumber(maximumTxPower)-1
-                	else
-                        	t = cmd('iwinfo ' .. interface50 .. ' txpowerlist | grep -n "19 dBm" | cut -f1 -d\':\'')
-       	                	maximumTxPower = string.gsub(t, "\n", "")
-       	                	maximumTxPower = tonumber(maximumTxPower)-1
-       	        	end
-       	        	--- set values (2nd pass)
-	                uci:set('wireless', interface50, 'txpower', maximumTxPower)
-                else -- outdoors
-        	        uci:set('wireless', interface50, 'country', country)
-	               	uci:delete('wireless', interface50, 'txpower')
-		end 
-                uci:save('wireless')
-                uci:commit('wireless')
+                end
+	end
+end
+
+--- restart with with new ht modes
+if interfac24 or interface50 then
+	t = cmd('/sbin/wifi reconfigure')
+	t = cmd('/sbin/wifi down')
+	t = cmd('kill -9  $(ps |grep /usr/sbin/hostapd|grep -v grep|awk \'{print $1}\')')
+	t = cmd('killall hostapd >/dev/null 2>&1')
+	t = cmd('rm -f /var/run/wifi-*.pid >/dev/null 2>&1')
+	t = cmd('/sbin/wifi up')
+	t = cmd('sleep 7')
+end
+
+--- 2.4G 
+if interface24 then
+	--- get maximum available power and step for 2.4G
+        t = cmd('iwinfo ' .. interface24 .. ' txpowerlist|grep mW|tr -d \'*\'|tail -n 1| awk \'{print $1}\'')
+        maximumTxPowerDb = string.gsub(t, "\n", "")
+        maximumTxPowerDb = tonumber(maximumTxPowerDb)
+	--- set values (2nd pass)
+        uci:set('wireless', interface24, 'txpower', maximumTxPowerDb)
+end
+
+--- 5G
+if interface50 then
+ 	if channel50 ~= 'auto' and uci:get('gluon', 'wireless', 'outdoor') ~= '1' then  --- do't do if outdoor is enabled
+	     	--- get maximum available power and step
+      		t = cmd('iwinfo ' .. interface50 .. ' txpowerlist|grep mW|tr -d \'*\'|tail -n 1| awk \'{print $1}\'')
+	      	maximumTxPowerDb = string.gsub(t, "\n", "")
+       		maximumTxPowerDb = tonumber(maximumTxPowerDb)
+	      	--- set values (2nd pass)
+	        uci:set('wireless', interface50, 'txpower', maximumTxPowerDb)
+	else -- outdoors
+        	uci:set('wireless', interface50, 'country', country)
+		uci:delete('wireless', interface50, 'txpower')
         end
-	if setupmode  == '0' then
+end
+
+--- restart stuff
+if interfac24 or interface50 then
+        uci:save('wireless')
+        uci:commit('wireless')
+	if setupmode  == '0' then --- restart wifi in case of full operation
 	        t = cmd('/sbin/wifi down')
         	t = cmd('kill -9  $(ps |grep /usr/sbin/hostapd|grep -v grep|awk \'{print $1}\')')
 	        t = cmd('killall hostapd >/dev/null 2>&1')
